requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
frozenlist==1.7.0
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.4.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.9
langchain-core==0.2.43
langchain-google-genai==1.0.7
langchain-postgres==0.0.14
langchain-text-splitters==0.2.4
langsmith==0.1.147
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.5.0
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.10.18
packaging==24.2
pdfkit==1.0.0
pgvector==0.3.6
pillow==11.1.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
pypdf==4.3.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.46.0
tenacity==8.5.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
yarl==1.20.1
